refactor(dashboard): split thread_api into a threads package
thread_api.py had grown to 2849 lines covering six unrelated concerns.
It is now agent/dashboard/threads/, one module per concern, largest 662
lines. thread_pins.py moves in as threads/pins.py; its only importer is
threads/listing.py.

The recovery-patch generator was a ~130-line Python program embedded as a
raw string, invisible to ruff, ty and pytest. It is now a real module at
agent/resources/recovery_patch.py, read as text and delivered by the same
heredoc. Making it visible immediately surfaced a lint error (UP022),
fixed here.

Drops four functions with no callers: _set_command_last_message_content,
_prefix_message_content, _prepend_message_content_block and
_latest_run_status.

Tests patched module globals as a flat namespace via
monkeypatch.setattr(thread_api, ...), which only worked while one module
held every binding. tests/conftest.py gains patch_thread_module, which
rebinds a name across every thread module that has it -- the same blast
radius those patches already had -- and raises if no module defines it.
