+++ agent/dashboard/threads/__init__.py
@@ -1,0 +1,5 @@
+"""Dashboard thread endpoints, split by concern.
+
+Import the submodule you need; nothing is re-exported here so that the
+dependency direction between these modules stays visible at each call site.
+"""
