+++ agent/utils/thread_ops.py
@@ -4,7 +4,7 @@ The webhook triggers (Slack / Linear / GitHub) dispatch through
 ``agent.dispatch.dispatch_agent_run`` with ``multitask_strategy="interrupt"``,
 so they no longer need a busy-check or an in-process lock. The store-queue
 below is retained for the dashboard's deliberate "inject a follow-up into a
-run that's already in flight" path (``thread_api.send_dashboard_message``).
+run that's already in flight" path (``threads.api.send_dashboard_message``).
 """
 
 import logging
