+++ tests/agent/test_plan_mode.py
@@ -4,8 +4,10 @@ from typing import Any
 import pytest
 
 from agent import server
-from agent.dashboard import thread_api
+from agent.dashboard.threads import runs as thread_runs
+from agent.dashboard.threads import summary as thread_summary
 from agent.prompt import construct_system_prompt
+from tests.conftest import patch_thread_module
 
 
 @pytest.mark.parametrize("enabled", [False, True])
@@ -117,10 +119,10 @@ def dashboard_run_client(monkeypatch: pytest.MonkeyPatch) -> _FakeLangGraphClien
     async def fake_resolve_email(login: str, profile: dict[str, Any]) -> str:
         return "octo@example.com"
 
-    monkeypatch.setattr(thread_api, "langgraph_client", lambda: client)
-    monkeypatch.setattr(thread_api, "get_profile", fake_get_profile)
-    monkeypatch.setattr(thread_api, "_ensure_dashboard_github_token", fake_ensure_token)
-    monkeypatch.setattr(thread_api, "_resolve_run_email", fake_resolve_email)
+    patch_thread_module(monkeypatch, "langgraph_client", lambda: client)
+    patch_thread_module(monkeypatch, "get_profile", fake_get_profile)
+    patch_thread_module(monkeypatch, "_ensure_dashboard_github_token", fake_ensure_token)
+    patch_thread_module(monkeypatch, "_resolve_run_email", fake_resolve_email)
     return client
 
 
@@ -141,7 +143,7 @@ def test_run_start_passes_plan_mode_when_enabled(
     dashboard_run_client: _FakeLangGraphClient,
 ) -> None:
     enriched = asyncio.run(
-        thread_api._enrich_run_start_command(
+        thread_runs._enrich_run_start_command(
             "thread-id",
             "octo",
             _run_start_command(True),
@@ -158,7 +160,7 @@ def test_run_start_omits_plan_mode_when_disabled(
     dashboard_run_client: _FakeLangGraphClient,
 ) -> None:
     enriched = asyncio.run(
-        thread_api._enrich_run_start_command(
+        thread_runs._enrich_run_start_command(
             "thread-id",
             "octo",
             _run_start_command(None),
@@ -172,12 +174,12 @@ def test_run_start_omits_plan_mode_when_disabled(
 
 
 async def test_thread_summary_reports_plan_mode() -> None:
-    summary = await thread_api._thread_summary(
+    summary = await thread_summary._thread_summary(
         {"thread_id": "t1", "metadata": {"source": "dashboard", "plan_mode": True}}
     )
     assert summary["planMode"] is True
 
-    summary_off = await thread_api._thread_summary(
+    summary_off = await thread_summary._thread_summary(
         {"thread_id": "t2", "metadata": {"source": "dashboard"}}
     )
     assert summary_off["planMode"] is False
