+++ tests/conftest.py
@@ -2,13 +2,30 @@
 
 from collections.abc import Iterator, Sequence
 from pathlib import Path
+from types import ModuleType
 from typing import Any
 
 import pytest
 
 from agent import store as agent_store
+from agent.dashboard.threads import access, api, diffs, listing, proxy, runs, summary
 from agent.utils import ttl_cache
 from agent.webhooks import common as webhook_common
+
+_THREAD_MODULES: tuple[ModuleType, ...] = (access, api, diffs, listing, proxy, runs, summary)
+
+
+def patch_thread_module(monkeypatch: pytest.MonkeyPatch, name: str, value: Any) -> None:
+    """Rebind ``name`` in every dashboard thread module that imports it.
+
+    The thread endpoints are split across modules that each hold their own
+    binding, so patching one would leave the others pointing at the real thing.
+    """
+    modules = [module for module in _THREAD_MODULES if hasattr(module, name)]
+    if not modules:
+        raise AttributeError(f"no dashboard thread module defines {name!r}")
+    for module in modules:
+        monkeypatch.setattr(module, name, value)
 
 
 class _FakeStoreNotFoundError(Exception):
