+++ tests/dashboard/test_dashboard_repo_optional.py
@@ -3,7 +3,9 @@ from typing import Any
 
 import pytest
 
-from agent.dashboard import thread_api
+from agent.dashboard.threads import runs as thread_runs
+from agent.dashboard.threads import summary as thread_summary
+from tests.conftest import patch_thread_module
 
 
 async def _fake_trace_url(thread_id: str, **kwargs: object) -> str:
@@ -11,17 +13,17 @@ async def _fake_trace_url(thread_id: str, **kwargs: object) -> str:
 
 
 def test_resolve_repo_config_parses_request_repo() -> None:
-    assert thread_api._resolve_repo_config("octo/repo") == {"owner": "octo", "name": "repo"}
+    assert thread_runs._resolve_repo_config("octo/repo") == {"owner": "octo", "name": "repo"}
 
 
 def test_resolve_repo_config_returns_empty_when_no_repo_given() -> None:
-    assert thread_api._resolve_repo_config(None) == {}
-    assert thread_api._resolve_repo_config("") == {}
-    assert thread_api._resolve_repo_config("not-a-repo") == {}
+    assert thread_runs._resolve_repo_config(None) == {}
+    assert thread_runs._resolve_repo_config("") == {}
+    assert thread_runs._resolve_repo_config("not-a-repo") == {}
 
 
 async def test_thread_summary_blanks_repo_when_absent() -> None:
-    summary = await thread_api._thread_summary(
+    summary = await thread_summary._thread_summary(
         {"thread_id": "t1", "metadata": {"source": "dashboard", "title": "no repo run"}}
     )
     assert summary["repo"] == ""
@@ -29,7 +31,7 @@ async def test_thread_summary_blanks_repo_when_absent() -> None:
 
 
 async def test_thread_summary_keeps_repo_when_present() -> None:
-    summary = await thread_api._thread_summary(
+    summary = await thread_summary._thread_summary(
         {
             "thread_id": "t2",
             "metadata": {
@@ -45,7 +47,7 @@ async def test_thread_summary_keeps_repo_when_present() -> None:
 
 
 async def test_thread_summary_classifies_legacy_schedule_metadata() -> None:
-    summary = await thread_api._thread_summary(
+    summary = await thread_summary._thread_summary(
         {
             "thread_id": "scheduled",
             "metadata": {
@@ -64,7 +66,7 @@ async def test_thread_summary_classifies_legacy_schedule_metadata() -> None:
 
 
 async def test_thread_summary_marks_automation_actions_posted_to_slack() -> None:
-    summary = await thread_api._thread_summary(
+    summary = await thread_summary._thread_summary(
         {
             "thread_id": "scheduled-action",
             "metadata": {
@@ -79,7 +81,7 @@ async def test_thread_summary_marks_automation_actions_posted_to_slack() -> None
 
 
 async def test_thread_summary_derives_issue_category_from_source_context() -> None:
-    summary = await thread_api._thread_summary(
+    summary = await thread_summary._thread_summary(
         {
             "thread_id": "linear",
             "metadata": {
@@ -94,12 +96,8 @@ async def test_thread_summary_derives_issue_category_from_source_context() -> No
 
 
 async def test_thread_summary_includes_trace_url(monkeypatch: pytest.MonkeyPatch) -> None:
-    monkeypatch.setattr(
-        thread_api,
-        "get_langsmith_trace_url",
-        _fake_trace_url,
-    )
-    summary = await thread_api._thread_summary(
+    patch_thread_module(monkeypatch, "get_langsmith_trace_url", _fake_trace_url)
+    summary = await thread_summary._thread_summary(
         {"thread_id": "t3", "metadata": {"source": "dashboard", "title": "traced run"}}
     )
     assert summary["traceUrl"] == "https://smith.example/t/t3"
@@ -156,10 +154,10 @@ def dashboard_run_client(monkeypatch: pytest.MonkeyPatch) -> _FakeLangGraphClien
     async def fake_resolve_email(login: str, profile: dict[str, Any]) -> str:
         return "octo@example.com"
 
-    monkeypatch.setattr(thread_api, "langgraph_client", lambda: client)
-    monkeypatch.setattr(thread_api, "get_profile", fake_get_profile)
-    monkeypatch.setattr(thread_api, "_ensure_dashboard_github_token", fake_ensure_token)
-    monkeypatch.setattr(thread_api, "_resolve_run_email", fake_resolve_email)
+    patch_thread_module(monkeypatch, "langgraph_client", lambda: client)
+    patch_thread_module(monkeypatch, "get_profile", fake_get_profile)
+    patch_thread_module(monkeypatch, "_ensure_dashboard_github_token", fake_ensure_token)
+    patch_thread_module(monkeypatch, "_resolve_run_email", fake_resolve_email)
     return client
 
 
@@ -171,7 +169,7 @@ def test_create_thread_record_omits_repo_less_marker_when_repo_unset(
     # ``_build_dashboard_configurable``. The thread record must not persist a
     # repo-less marker when the repo is simply unset (not explicitly cleared).
     asyncio.run(
-        thread_api._create_dashboard_thread_record(
+        thread_runs._create_dashboard_thread_record(
             "thread-id",
             login="octo",
             repo_config={},
@@ -180,7 +178,7 @@ def test_create_thread_record_omits_repo_less_marker_when_repo_unset(
     )
 
     configurable = asyncio.run(
-        thread_api._build_dashboard_configurable("thread-id", "octo", {"source": "dashboard"})
+        thread_runs._build_dashboard_configurable("thread-id", "octo", {"source": "dashboard"})
     )
     assert "repo_explicitly_none" not in configurable
     assert "repo" not in configurable
@@ -190,7 +188,7 @@ def test_build_configurable_marks_repo_less_config_when_explicit(
     dashboard_run_client: _FakeLangGraphClient,
 ) -> None:
     configurable = asyncio.run(
-        thread_api._build_dashboard_configurable(
+        thread_runs._build_dashboard_configurable(
             "thread-id",
             "octo",
             {"source": "dashboard", "repo_explicitly_none": True},
@@ -204,7 +202,7 @@ def test_build_configurable_includes_repo_when_configured(
     dashboard_run_client: _FakeLangGraphClient,
 ) -> None:
     configurable = asyncio.run(
-        thread_api._build_dashboard_configurable(
+        thread_runs._build_dashboard_configurable(
             "thread-id",
             "octo",
             {"source": "dashboard", "repo_owner": "octo", "repo_name": "repo"},
