+++ tests/dashboard/test_dashboard_run_email.py
@@ -3,8 +3,8 @@ from typing import Any
 import pytest
 
 from agent import store as agent_store
-from agent.dashboard import thread_api
 from agent.dashboard import user_mappings as um
+from agent.dashboard.threads import access as thread_access
 
 
 class _FakeStore:
@@ -52,10 +52,12 @@ async def test_run_email_prefers_github_mapping(fake_store: _FakeStore) -> None:
     )
     # OAuth profile carries a personal account that isn't an org member.
     profile = {"email": "johannesduplessis117@gmail.com"}
-    assert await thread_api._resolve_run_email("johannes117", profile) == "johannes@langchain.dev"
+    assert (
+        await thread_access._resolve_run_email("johannes117", profile) == "johannes@langchain.dev"
+    )
 
 
 @pytest.mark.asyncio
 async def test_run_email_falls_back_to_profile_when_unmapped(fake_store: _FakeStore) -> None:
     profile = {"email": "someone@example.com"}
-    assert await thread_api._resolve_run_email("nomapping", profile) == "someone@example.com"
+    assert await thread_access._resolve_run_email("nomapping", profile) == "someone@example.com"
