+++ tests/dashboard/test_dashboard_thread_api.py
@@ -9,10 +9,17 @@ from xml.etree import ElementTree
 import pytest
 from fastapi import HTTPException
 
-from agent.dashboard import routes, thread_api
+from agent.dashboard import routes
 from agent.dashboard.agent_overrides import resolve_agent_model_id
 from agent.dashboard.options import model_supports_images
+from agent.dashboard.threads import api as thread_api
+from agent.dashboard.threads import diffs as thread_diffs
+from agent.dashboard.threads import listing as thread_listing
+from agent.dashboard.threads import proxy as thread_proxy
+from agent.dashboard.threads import runs as thread_runs
+from agent.dashboard.threads import summary as thread_summary
 from agent.dashboard.ttft import AssistantTextObservation
+from tests.conftest import patch_thread_module
 
 _TEXT_ONLY_MODEL = "fireworks:accounts/fireworks/models/deepseek-v4-pro"
 _VISION_MODEL = "openai:gpt-5.6-sol"
@@ -30,7 +37,7 @@ def _empty_thread_pins(monkeypatch) -> None:
     async def empty_pins(login: str) -> list[str]:
         return []
 
-    monkeypatch.setattr(thread_api, "list_thread_pin_ids", empty_pins)
+    patch_thread_module(monkeypatch, "list_thread_pin_ids", empty_pins)
 
 
 async def test_rename_thread_trims_title_and_clears_seed(monkeypatch) -> None:
@@ -38,19 +45,19 @@ async def test_rename_thread_trims_title_and_clears_seed(monkeypatch) -> None:
     thread = {"thread_id": "thread-1", "metadata": metadata}
     authorized = AsyncMock(return_value=thread)
     update = AsyncMock()
-    monkeypatch.setattr(thread_api, "_authorized_thread", authorized)
-    monkeypatch.setattr(
-        thread_api, "_thread_summary", AsyncMock(side_effect=lambda t: t["metadata"])
+    patch_thread_module(monkeypatch, "_authorized_thread", authorized)
+    patch_thread_module(
+        monkeypatch, "_thread_summary", AsyncMock(side_effect=lambda t: t["metadata"])
     )
-    monkeypatch.setattr(
-        thread_api,
+    patch_thread_module(
+        monkeypatch,
         "langgraph_client",
         lambda: SimpleNamespace(threads=SimpleNamespace(update=update)),
     )
 
     result = await routes.api_rename_thread(
         "thread-1",
-        thread_api.ThreadRenameBody(title="  New title  "),
+        thread_runs.ThreadRenameBody(title="  New title  "),
         {"sub": "alice", "email": "alice@example.com"},
     )
 
@@ -61,8 +68,8 @@ async def test_rename_thread_trims_title_and_clears_seed(monkeypatch) -> None:
     assert result == {**metadata, "title": "New title", "title_seed": None}
 
 
-def _image() -> thread_api.DashboardImageBody:
-    return thread_api.DashboardImageBody(
+def _image() -> thread_runs.DashboardImageBody:
+    return thread_runs.DashboardImageBody(
         base64=base64.b64encode(b"image").decode("ascii"),
         mimeType="image/png",
     )
@@ -75,14 +82,14 @@ def test_model_supports_images_marks_text_only_fireworks_models() -> None:
 
 def test_user_message_content_rejects_images_for_text_only_model() -> None:
     with pytest.raises(HTTPException) as exc_info:
-        thread_api._user_message_content("see attached", [_image()], model_id=_TEXT_ONLY_MODEL)
+        thread_runs._user_message_content("see attached", [_image()], model_id=_TEXT_ONLY_MODEL)
 
     assert exc_info.value.status_code == 422
     assert "does not support image input" in exc_info.value.detail
 
 
 def test_user_message_content_allows_images_for_vision_model() -> None:
-    content = thread_api._user_message_content("see attached", [_image()], model_id=_VISION_MODEL)
+    content = thread_runs._user_message_content("see attached", [_image()], model_id=_VISION_MODEL)
 
     assert isinstance(content, list)
     assert content[-1] == {"type": "text", "text": "see attached"}
@@ -92,7 +99,7 @@ def test_user_message_content_allows_images_for_vision_model() -> None:
 def test_langgraph_proxy_headers_include_api_key(monkeypatch) -> None:
     monkeypatch.setenv("LANGSMITH_API_KEY", "ls-key")
 
-    headers = thread_api._langgraph_proxy_headers(accept="text/event-stream")
+    headers = thread_proxy._langgraph_proxy_headers(accept="text/event-stream")
 
     assert headers["X-API-Key"] == "ls-key"
     assert headers["Accept"] == "text/event-stream"
@@ -103,9 +110,9 @@ async def test_resolve_agent_model_choice_applies_profile_before_team_default(mo
         assert role == "agent"
         return _VISION_MODEL, "medium"
 
-    monkeypatch.setattr(thread_api, "get_team_default_model", fake_team_default)
+    patch_thread_module(monkeypatch, "get_team_default_model", fake_team_default)
 
-    model_id, effort = await thread_api._resolve_agent_model_choice(
+    model_id, effort = await thread_runs._resolve_agent_model_choice(
         {"default_model": _TEXT_ONLY_MODEL, "reasoning_effort": "high"},
         None,
         None,
@@ -119,9 +126,9 @@ async def test_resolve_agent_model_choice_applies_request_before_profile(monkeyp
         assert role == "agent"
         return _VISION_MODEL, "medium"
 
-    monkeypatch.setattr(thread_api, "get_team_default_model", fake_team_default)
+    patch_thread_module(monkeypatch, "get_team_default_model", fake_team_default)
 
-    model_id, effort = await thread_api._resolve_agent_model_choice(
+    model_id, effort = await thread_runs._resolve_agent_model_choice(
         {"default_model": _TEXT_ONLY_MODEL, "reasoning_effort": "high"},
         "anthropic:claude-opus-5",
         "high",
@@ -134,9 +141,9 @@ async def test_resolve_agent_model_choice_deprecated_request_uses_team_default(m
     async def fake_team_default(role: str) -> tuple[str, str]:
         return _VISION_MODEL, "medium"
 
-    monkeypatch.setattr(thread_api, "get_team_default_model", fake_team_default)
+    patch_thread_module(monkeypatch, "get_team_default_model", fake_team_default)
 
-    model_id, effort = await thread_api._resolve_agent_model_choice(
+    model_id, effort = await thread_runs._resolve_agent_model_choice(
         {"default_model": "anthropic:claude-opus-5", "reasoning_effort": "high"},
         "fireworks:accounts/fireworks/models/glm-5p2",
         "high",
@@ -231,16 +238,16 @@ def _patch_new_thread_deps(monkeypatch, *, profile: dict[str, object]) -> None:
     async def fake_resolve_email(login: str, prof: dict[str, object]) -> str:
         return f"{login}@example.com"
 
-    monkeypatch.setattr(thread_api, "get_profile", fake_profile)
-    monkeypatch.setattr(thread_api, "get_team_default_model", fake_team_default)
-    monkeypatch.setattr(thread_api, "_ensure_dashboard_github_token", fake_ensure_token)
-    monkeypatch.setattr(thread_api, "_resolve_run_email", fake_resolve_email)
+    patch_thread_module(monkeypatch, "get_profile", fake_profile)
+    patch_thread_module(monkeypatch, "get_team_default_model", fake_team_default)
+    patch_thread_module(monkeypatch, "_ensure_dashboard_github_token", fake_ensure_token)
+    patch_thread_module(monkeypatch, "_resolve_run_email", fake_resolve_email)
 
 
 async def test_enrich_run_start_command_creates_and_stamps_new_thread(monkeypatch) -> None:
     created: dict[str, object] = {}
     _patch_new_thread_deps(monkeypatch, profile={})
-    monkeypatch.setattr(thread_api, "langgraph_client", lambda: _new_thread_client(created))
+    patch_thread_module(monkeypatch, "langgraph_client", lambda: _new_thread_client(created))
 
     command = {
         "method": "run.start",
@@ -256,7 +263,7 @@ async def test_enrich_run_start_command_creates_and_stamps_new_thread(monkeypatc
         },
     }
 
-    enriched = await thread_api._enrich_run_start_command(
+    enriched = await thread_runs._enrich_run_start_command(
         "new-tid",
         "octocat",
         command,
@@ -301,7 +308,7 @@ async def test_enrich_run_start_command_uses_vision_fallback_for_text_only_model
         monkeypatch,
         profile={"default_model": _TEXT_ONLY_MODEL, "reasoning_effort": "high"},
     )
-    monkeypatch.setattr(thread_api, "langgraph_client", lambda: _new_thread_client(created))
+    patch_thread_module(monkeypatch, "langgraph_client", lambda: _new_thread_client(created))
 
     image = _image()
     command = {
@@ -326,7 +333,7 @@ async def test_enrich_run_start_command_uses_vision_fallback_for_text_only_model
         },
     }
 
-    enriched = await thread_api._enrich_run_start_command(
+    enriched = await thread_runs._enrich_run_start_command(
         "new-tid",
         "octocat",
         command,
@@ -350,17 +357,17 @@ def _thread_with_metadata(metadata: dict) -> dict:
 
 
 async def test_thread_summary_exposes_attention_reason() -> None:
-    flagged = await thread_api._thread_summary(
+    flagged = await thread_summary._thread_summary(
         _thread_with_metadata({"title": "Ship it", "attention_reason": "prs_closed"})
     )
-    quiet = await thread_api._thread_summary(_thread_with_metadata({"title": "Ship it"}))
+    quiet = await thread_summary._thread_summary(_thread_with_metadata({"title": "Ship it"}))
 
     assert flagged["attentionReason"] == "prs_closed"
     assert quiet["attentionReason"] is None
 
 
 async def test_thread_summary_includes_pr_and_diff_stats() -> None:
-    summary = await thread_api._thread_summary(
+    summary = await thread_summary._thread_summary(
         _thread_with_metadata(
             {
                 "repo_full_name": "langchain-ai/open-swe",
@@ -389,7 +396,7 @@ async def test_thread_summary_includes_pr_and_diff_stats() -> None:
 
 
 async def test_thread_summary_includes_pull_requests_across_repositories() -> None:
-    summary = await thread_api._thread_summary(
+    summary = await thread_summary._thread_summary(
         _thread_with_metadata(
             {
                 "repo_full_name": "langchain-ai/open-swe",
@@ -445,7 +452,7 @@ async def test_thread_summary_uses_configured_repo_for_display() -> None:
         "working_repo_full_name": "observed/checkout",
     }
 
-    summary = await thread_api._thread_summary(_thread_with_metadata(metadata))
+    summary = await thread_summary._thread_summary(_thread_with_metadata(metadata))
 
     assert summary["repo"] == "default"
     assert summary["repoFullName"] == "trusted/default"
@@ -454,7 +461,7 @@ async def test_thread_summary_uses_configured_repo_for_display() -> None:
 
 
 async def test_thread_summary_defaults_unknown_pr_state_to_open() -> None:
-    summary = await thread_api._thread_summary(
+    summary = await thread_summary._thread_summary(
         _thread_with_metadata(
             {
                 "pr_number": 7,
@@ -468,20 +475,22 @@ async def test_thread_summary_defaults_unknown_pr_state_to_open() -> None:
 
 
 async def test_thread_summary_omits_pr_when_no_pr_metadata() -> None:
-    summary = await thread_api._thread_summary(_thread_with_metadata({"title": "No PR"}))
+    summary = await thread_summary._thread_summary(_thread_with_metadata({"title": "No PR"}))
 
     assert "pr" not in summary
     assert "diffStats" not in summary
 
 
 async def test_thread_summary_exposes_sandbox_id() -> None:
-    summary = await thread_api._thread_summary(_thread_with_metadata({"sandbox_id": "sb-abc123"}))
+    summary = await thread_summary._thread_summary(
+        _thread_with_metadata({"sandbox_id": "sb-abc123"})
+    )
 
     assert summary["sandboxId"] == "sb-abc123"
 
 
 async def test_thread_summary_hides_creating_sandbox_sentinel() -> None:
-    summary = await thread_api._thread_summary(
+    summary = await thread_summary._thread_summary(
         _thread_with_metadata({"sandbox_id": "__creating__"})
     )
 
@@ -502,7 +511,7 @@ async def test_terminal_sandbox_requires_existing_sandbox(monkeypatch) -> None:
     class FakeClient:
         threads = FakeThreads()
 
-    monkeypatch.setattr(thread_api, "langgraph_client", lambda: FakeClient())
+    patch_thread_module(monkeypatch, "langgraph_client", lambda: FakeClient())
 
     assert await thread_api.get_dashboard_terminal_sandbox("tid", "teammate") == (
         "sandbox-123",
@@ -516,8 +525,8 @@ async def test_terminal_sandbox_requires_existing_sandbox(monkeypatch) -> None:
 
 
 async def test_thread_summary_includes_slack_source_urls_for_private_repo(monkeypatch) -> None:
-    monkeypatch.setattr(thread_api, "SLACK_TEAM_ID", "T workspace")
-    summary = await thread_api._thread_summary(
+    patch_thread_module(monkeypatch, "SLACK_TEAM_ID", "T workspace")
+    summary = await thread_summary._thread_summary(
         _thread_with_metadata(
             {
                 "source": "slack",
@@ -540,8 +549,8 @@ async def test_thread_summary_includes_slack_source_urls_for_private_repo(monkey
 
 
 async def test_thread_summary_includes_slack_source_urls_for_public_repo(monkeypatch) -> None:
-    monkeypatch.setattr(thread_api, "SLACK_TEAM_ID", "T-workspace")
-    summary = await thread_api._thread_summary(
+    patch_thread_module(monkeypatch, "SLACK_TEAM_ID", "T-workspace")
+    summary = await thread_summary._thread_summary(
         _thread_with_metadata(
             {
                 "source": "slack",
@@ -564,8 +573,8 @@ async def test_thread_summary_includes_slack_source_urls_for_public_repo(monkeyp
 
 
 async def test_thread_summary_includes_code_channel_url(monkeypatch) -> None:
-    monkeypatch.setattr(thread_api, "SLACK_TEAM_ID", "T team&workspace")
-    summary = await thread_api._thread_summary(
+    patch_thread_module(monkeypatch, "SLACK_TEAM_ID", "T team&workspace")
+    summary = await thread_summary._thread_summary(
         _thread_with_metadata(
             {
                 "source": "slack",
@@ -582,7 +591,7 @@ async def test_thread_summary_includes_code_channel_url(monkeypatch) -> None:
 
 
 async def test_thread_summary_omits_code_channel_url_without_team() -> None:
-    summary = await thread_api._thread_summary(
+    summary = await thread_summary._thread_summary(
         _thread_with_metadata(
             {
                 "source": "slack",
@@ -595,7 +604,7 @@ async def test_thread_summary_omits_code_channel_url_without_team() -> None:
 
 
 async def test_thread_summary_omits_code_channel_url_for_slack_thread() -> None:
-    summary = await thread_api._thread_summary(
+    summary = await thread_summary._thread_summary(
         _thread_with_metadata(
             {
                 "source": "slack",
@@ -617,10 +626,10 @@ async def test_recovery_patch_reaches_any_authenticated_user(monkeypatch) -> Non
     class FakeClient:
         threads = FakeThreads()
 
-    monkeypatch.setattr(thread_api, "langgraph_client", lambda: FakeClient())
+    patch_thread_module(monkeypatch, "langgraph_client", lambda: FakeClient())
 
     with pytest.raises(HTTPException) as exc_info:
-        await thread_api.get_dashboard_thread_recovery_patch("tid", "teammate")
+        await thread_diffs.get_dashboard_thread_recovery_patch("tid", "teammate")
 
     assert exc_info.value.status_code == 404
     assert "sandbox" in exc_info.value.detail
@@ -630,10 +639,10 @@ async def test_recovery_patch_requires_sandbox(monkeypatch) -> None:
     async def fake_authorized_thread(thread_id: str, login: str, *, email: str | None = None):
         return {"thread_id": thread_id, "metadata": {"source": "dashboard", "github_login": login}}
 
-    monkeypatch.setattr(thread_api, "_authorized_thread", fake_authorized_thread)
+    patch_thread_module(monkeypatch, "_authorized_thread", fake_authorized_thread)
 
     with pytest.raises(HTTPException) as exc_info:
-        await thread_api.get_dashboard_thread_recovery_patch("tid", "octocat")
+        await thread_diffs.get_dashboard_thread_recovery_patch("tid", "octocat")
 
     assert exc_info.value.status_code == 404
     assert "sandbox" in exc_info.value.detail
@@ -656,7 +665,7 @@ async def test_recovery_patch_downloads_generated_patch(monkeypatch) -> None:
     class FakeSandbox:
         async def aexecute(self, command: str, *, timeout: int | None = None):
             assert "repo" in command
-            assert timeout == thread_api._RECOVERY_PATCH_TIMEOUT_SECONDS
+            assert timeout == thread_diffs._RECOVERY_PATCH_TIMEOUT_SECONDS
             return SimpleNamespace(
                 output=json.dumps({"ok": True, "path": "/tmp/open-swe-tid.patch", "size": 11}),
                 exit_code=0,
@@ -666,10 +675,10 @@ async def test_recovery_patch_downloads_generated_patch(monkeypatch) -> None:
             assert paths == ["/tmp/open-swe-tid.patch"]
             return [SimpleNamespace(content=b"patch bytes")]
 
-    monkeypatch.setattr(thread_api, "_authorized_thread", fake_authorized_thread)
-    monkeypatch.setattr(thread_api, "create_sandbox", AsyncMock(return_value=FakeSandbox()))
+    patch_thread_module(monkeypatch, "_authorized_thread", fake_authorized_thread)
+    patch_thread_module(monkeypatch, "create_sandbox", AsyncMock(return_value=FakeSandbox()))
 
-    content, filename = await thread_api.get_dashboard_thread_recovery_patch("tid", "octocat")
+    content, filename = await thread_diffs.get_dashboard_thread_recovery_patch("tid", "octocat")
 
     assert content == b"patch bytes"
     assert filename == "open-swe-tid.patch"
@@ -686,11 +695,11 @@ async def test_recovery_patch_rejects_empty_patch(monkeypatch) -> None:
                 exit_code=0,
             )
 
-    monkeypatch.setattr(thread_api, "_authorized_thread", fake_authorized_thread)
-    monkeypatch.setattr(thread_api, "create_sandbox", AsyncMock(return_value=FakeSandbox()))
+    patch_thread_module(monkeypatch, "_authorized_thread", fake_authorized_thread)
+    patch_thread_module(monkeypatch, "create_sandbox", AsyncMock(return_value=FakeSandbox()))
 
     with pytest.raises(HTTPException) as exc_info:
-        await thread_api.get_dashboard_thread_recovery_patch("tid", "octocat")
+        await thread_diffs.get_dashboard_thread_recovery_patch("tid", "octocat")
 
     assert exc_info.value.status_code == 404
     assert "changes" in exc_info.value.detail
@@ -707,29 +716,28 @@ async def test_recovery_patch_enforces_size_limit(monkeypatch) -> None:
                     {
                         "ok": True,
                         "path": "/tmp/open-swe-tid.patch",
-                        "size": thread_api._RECOVERY_PATCH_LIMIT_BYTES + 1,
+                        "size": thread_diffs._RECOVERY_PATCH_LIMIT_BYTES + 1,
                     }
                 ),
                 exit_code=0,
             )
 
-    monkeypatch.setattr(thread_api, "_authorized_thread", fake_authorized_thread)
-    monkeypatch.setattr(thread_api, "create_sandbox", AsyncMock(return_value=FakeSandbox()))
+    patch_thread_module(monkeypatch, "_authorized_thread", fake_authorized_thread)
+    patch_thread_module(monkeypatch, "create_sandbox", AsyncMock(return_value=FakeSandbox()))
 
     with pytest.raises(HTTPException) as exc_info:
-        await thread_api.get_dashboard_thread_recovery_patch("tid", "octocat")
+        await thread_diffs.get_dashboard_thread_recovery_patch("tid", "octocat")
 
     assert exc_info.value.status_code == 413
 
 
 def test_recovery_patch_searches_command_cwd_before_workspace_fallback() -> None:
-    command = thread_api._recovery_patch_command(
+    command = thread_diffs._recovery_patch_command(
         {"repo_name": "repo", "base_branch": "main"},
         "tid",
     )
 
-    assert "Path.cwd().resolve()" in command
-    assert "WORKSPACE_FALLBACK = Path('/workspace')" in command
+    assert 'WORKSPACE_FALLBACK = Path("/workspace")' in command
     assert "roots = [Path.cwd().resolve(), WORKSPACE_FALLBACK]" in command
 
 
@@ -743,11 +751,11 @@ async def test_proxy_commands_lazily_creates_missing_thread_only_for_run_start(
     class MissingClient:
         threads = MissingThreads()
 
-    monkeypatch.setattr(thread_api, "langgraph_client", lambda: MissingClient())
+    patch_thread_module(monkeypatch, "langgraph_client", lambda: MissingClient())
 
     # A non-run.start command against a thread that doesn't exist yet is a 404.
     with pytest.raises(HTTPException) as exc_info:
-        await thread_api.proxy_dashboard_thread_commands(
+        await thread_proxy.proxy_dashboard_thread_commands(
             "ghost", "octocat", b'{"method": "run.cancel"}'
         )
     assert exc_info.value.status_code == 404
@@ -773,17 +781,17 @@ async def test_enrich_run_start_command_attributes_non_owner_message(monkeypatch
     async def fake_resolve_email(login: str, profile: dict[str, object]) -> str:
         return f"{login}@example.com"
 
-    monkeypatch.setattr(thread_api, "langgraph_client", lambda: FakeClient())
-    monkeypatch.setattr(thread_api, "get_profile", fake_get_profile)
-    monkeypatch.setattr(thread_api, "_ensure_dashboard_github_token", fake_ensure_token)
-    monkeypatch.setattr(thread_api, "_resolve_run_email", fake_resolve_email)
+    patch_thread_module(monkeypatch, "langgraph_client", lambda: FakeClient())
+    patch_thread_module(monkeypatch, "get_profile", fake_get_profile)
+    patch_thread_module(monkeypatch, "_ensure_dashboard_github_token", fake_ensure_token)
+    patch_thread_module(monkeypatch, "_resolve_run_email", fake_resolve_email)
 
     command = {
         "method": "run.start",
         "params": {"input": {"messages": [{"role": "user", "content": "fix the bug"}]}},
     }
 
-    enriched = await thread_api._enrich_run_start_command(
+    enriched = await thread_runs._enrich_run_start_command(
         "tid",
         "teammate",
         command,
@@ -820,10 +828,10 @@ async def test_enrich_run_start_command_adds_web_handoff_for_slack_thread(monkey
     async def fake_resolve_email(login: str, profile: dict[str, object]) -> str:
         return f"{login}@example.com"
 
-    monkeypatch.setattr(thread_api, "langgraph_client", lambda: FakeClient())
-    monkeypatch.setattr(thread_api, "get_profile", fake_get_profile)
-    monkeypatch.setattr(thread_api, "_ensure_dashboard_github_token", fake_ensure_token)
-    monkeypatch.setattr(thread_api, "_resolve_run_email", fake_resolve_email)
+    patch_thread_module(monkeypatch, "langgraph_client", lambda: FakeClient())
+    patch_thread_module(monkeypatch, "get_profile", fake_get_profile)
+    patch_thread_module(monkeypatch, "_ensure_dashboard_github_token", fake_ensure_token)
+    patch_thread_module(monkeypatch, "_resolve_run_email", fake_resolve_email)
 
     command = {
         "method": "run.start",
@@ -834,7 +842,7 @@ async def test_enrich_run_start_command_adds_web_handoff_for_slack_thread(monkey
         },
     }
 
-    enriched = await thread_api._enrich_run_start_command(
+    enriched = await thread_runs._enrich_run_start_command(
         "tid",
         "teammate",
         command,
@@ -874,10 +882,10 @@ async def test_enrich_run_start_command_adds_web_handoff_before_image_blocks(mon
     async def fake_resolve_email(login: str, profile: dict[str, object]) -> str:
         return f"{login}@example.com"
 
-    monkeypatch.setattr(thread_api, "langgraph_client", lambda: FakeClient())
-    monkeypatch.setattr(thread_api, "get_profile", fake_get_profile)
-    monkeypatch.setattr(thread_api, "_ensure_dashboard_github_token", fake_ensure_token)
-    monkeypatch.setattr(thread_api, "_resolve_run_email", fake_resolve_email)
+    patch_thread_module(monkeypatch, "langgraph_client", lambda: FakeClient())
+    patch_thread_module(monkeypatch, "get_profile", fake_get_profile)
+    patch_thread_module(monkeypatch, "_ensure_dashboard_github_token", fake_ensure_token)
+    patch_thread_module(monkeypatch, "_resolve_run_email", fake_resolve_email)
 
     command = {
         "method": "run.start",
@@ -893,7 +901,7 @@ async def test_enrich_run_start_command_adds_web_handoff_before_image_blocks(mon
         },
     }
 
-    enriched = await thread_api._enrich_run_start_command(
+    enriched = await thread_runs._enrich_run_start_command(
         "tid",
         "teammate",
         command,
@@ -927,17 +935,17 @@ async def test_enrich_run_start_command_does_not_attribute_owner_message(monkeyp
     async def fake_resolve_email(login: str, profile: dict[str, object]) -> str:
         return f"{login}@example.com"
 
-    monkeypatch.setattr(thread_api, "langgraph_client", lambda: FakeClient())
-    monkeypatch.setattr(thread_api, "get_profile", fake_get_profile)
-    monkeypatch.setattr(thread_api, "_ensure_dashboard_github_token", fake_ensure_token)
-    monkeypatch.setattr(thread_api, "_resolve_run_email", fake_resolve_email)
+    patch_thread_module(monkeypatch, "langgraph_client", lambda: FakeClient())
+    patch_thread_module(monkeypatch, "get_profile", fake_get_profile)
+    patch_thread_module(monkeypatch, "_ensure_dashboard_github_token", fake_ensure_token)
+    patch_thread_module(monkeypatch, "_resolve_run_email", fake_resolve_email)
 
     command = {
         "method": "run.start",
         "params": {"input": {"messages": [{"role": "user", "content": "fix the bug"}]}},
     }
 
-    enriched = await thread_api._enrich_run_start_command(
+    enriched = await thread_runs._enrich_run_start_command(
         "tid",
         "owner",
         command,
@@ -972,10 +980,10 @@ async def test_enrich_run_start_command_allowlists_client_configurable(monkeypat
         assert login == "octocat"
         return "octocat@example.com"
 
-    monkeypatch.setattr(thread_api, "langgraph_client", lambda: FakeClient())
-    monkeypatch.setattr(thread_api, "get_profile", fake_get_profile)
-    monkeypatch.setattr(thread_api, "_ensure_dashboard_github_token", fake_ensure_token)
-    monkeypatch.setattr(thread_api, "_resolve_run_email", fake_resolve_email)
+    patch_thread_module(monkeypatch, "langgraph_client", lambda: FakeClient())
+    patch_thread_module(monkeypatch, "get_profile", fake_get_profile)
+    patch_thread_module(monkeypatch, "_ensure_dashboard_github_token", fake_ensure_token)
+    patch_thread_module(monkeypatch, "_resolve_run_email", fake_resolve_email)
 
     command = {
         "method": "run.start",
@@ -993,7 +1001,7 @@ async def test_enrich_run_start_command_allowlists_client_configurable(monkeypat
         },
     }
 
-    enriched = await thread_api._enrich_run_start_command(
+    enriched = await thread_runs._enrich_run_start_command(
         "tid",
         "octocat",
         command,
@@ -1081,17 +1089,17 @@ async def test_proxy_run_start_from_slack_thread_updates_trace_reply(monkeypatch
         }
         return True
 
-    monkeypatch.setattr(thread_api, "langgraph_client", lambda: FakeClient())
-    monkeypatch.setattr(thread_api, "get_profile", fake_get_profile)
-    monkeypatch.setattr(thread_api, "_ensure_dashboard_github_token", fake_ensure_token)
-    monkeypatch.setattr(thread_api, "_resolve_run_email", fake_resolve_email)
-    monkeypatch.setattr(thread_api, "_now_ms", lambda: 123_456)
-    monkeypatch.setattr(thread_api.httpx2, "AsyncClient", FakeAsyncClient)
-    monkeypatch.setattr(
-        thread_api, "update_slack_trace_reply_for_web_handoff", fake_update_trace_reply
+    patch_thread_module(monkeypatch, "langgraph_client", lambda: FakeClient())
+    patch_thread_module(monkeypatch, "get_profile", fake_get_profile)
+    patch_thread_module(monkeypatch, "_ensure_dashboard_github_token", fake_ensure_token)
+    patch_thread_module(monkeypatch, "_resolve_run_email", fake_resolve_email)
+    patch_thread_module(monkeypatch, "_now_ms", lambda: 123_456)
+    monkeypatch.setattr(thread_proxy.httpx2, "AsyncClient", FakeAsyncClient)
+    patch_thread_module(
+        monkeypatch, "update_slack_trace_reply_for_web_handoff", fake_update_trace_reply
     )
 
-    status, body, _ = await thread_api.proxy_dashboard_thread_commands(
+    status, body, _ = await thread_proxy.proxy_dashboard_thread_commands(
         "tid",
         "octocat",
         b'{"method":"run.start","params":{"input":{"messages":[{"role":"user","content":"continue here"}]}}}',
@@ -1194,10 +1202,10 @@ async def test_run_ttft_observer_records_first_assistant_text(
             return FakeStreamContext()
 
     record = AsyncMock()
-    monkeypatch.setattr(thread_api.httpx2, "AsyncClient", FakeAsyncClient)
-    monkeypatch.setattr(thread_api, "record_dashboard_thread_ttft", record)
+    monkeypatch.setattr(thread_proxy.httpx2, "AsyncClient", FakeAsyncClient)
+    patch_thread_module(monkeypatch, "record_dashboard_thread_ttft", record)
 
-    await thread_api._observe_dashboard_run_ttft("thread-1", "run-1", 1_000)
+    await thread_proxy._observe_dashboard_run_ttft("thread-1", "run-1", 1_000)
 
     record.assert_awaited_once_with(
         AssistantTextObservation(run_id="run-1", event_timestamp_ms=2_250),
@@ -1218,10 +1226,10 @@ async def test_proxy_commands_rejects_non_object_body(monkeypatch) -> None:
     class FakeClient:
         threads = FakeThreads()
 
-    monkeypatch.setattr(thread_api, "langgraph_client", lambda: FakeClient())
+    patch_thread_module(monkeypatch, "langgraph_client", lambda: FakeClient())
 
     with pytest.raises(HTTPException) as exc_info:
-        await thread_api.proxy_dashboard_thread_commands("tid", "octocat", b"[]")
+        await thread_proxy.proxy_dashboard_thread_commands("tid", "octocat", b"[]")
 
     assert exc_info.value.status_code == 400
 
@@ -1234,10 +1242,10 @@ async def test_proxy_commands_rejects_unsurfaced_thread(monkeypatch) -> None:
     class InternalClient:
         threads = InternalThreads()
 
-    monkeypatch.setattr(thread_api, "langgraph_client", lambda: InternalClient())
+    patch_thread_module(monkeypatch, "langgraph_client", lambda: InternalClient())
 
     with pytest.raises(HTTPException) as exc_info:
-        await thread_api.proxy_dashboard_thread_commands(
+        await thread_proxy.proxy_dashboard_thread_commands(
             "tid", "teammate", b'{"method": "input.respond"}'
         )
     assert exc_info.value.status_code == 404
@@ -1259,10 +1267,10 @@ async def test_proxy_commands_rejects_non_admin_on_admin_thread(monkeypatch) -> 
         threads = AdminThreads()
 
     monkeypatch.setenv("CONFIGURED_ADMINS", "workspace-admin")
-    monkeypatch.setattr(thread_api, "langgraph_client", lambda: AdminClient())
+    patch_thread_module(monkeypatch, "langgraph_client", lambda: AdminClient())
 
     with pytest.raises(HTTPException) as exc_info:
-        await thread_api.proxy_dashboard_thread_commands(
+        await thread_proxy.proxy_dashboard_thread_commands(
             "tid", "teammate", b'{"method": "run.start"}'
         )
 
@@ -1307,17 +1315,17 @@ async def test_proxy_commands_preserves_admin_writes_and_owner_reads(monkeypatch
             return FakeResponse()
 
     monkeypatch.setenv("CONFIGURED_ADMINS", "workspace-admin,another-admin")
-    monkeypatch.setattr(thread_api, "langgraph_client", lambda: AdminClient())
-    monkeypatch.setattr(thread_api.httpx2, "AsyncClient", FakeAsyncClient)
+    patch_thread_module(monkeypatch, "langgraph_client", lambda: AdminClient())
+    monkeypatch.setattr(thread_proxy.httpx2, "AsyncClient", FakeAsyncClient)
 
-    status_code, _, _ = await thread_api.proxy_dashboard_thread_commands(
+    status_code, _, _ = await thread_proxy.proxy_dashboard_thread_commands(
         "tid", "another-admin", b'{"method": "input.respond"}'
     )
 
     assert status_code == 200
 
     monkeypatch.setenv("CONFIGURED_ADMINS", "another-admin")
-    status_code, _, _ = await thread_api.proxy_dashboard_thread_commands(
+    status_code, _, _ = await thread_proxy.proxy_dashboard_thread_commands(
         "tid", "workspace-admin", b'{"method": "agent.getTree"}'
     )
 
@@ -1337,10 +1345,10 @@ async def test_run_cancel_rejects_unsurfaced_thread(monkeypatch) -> None:
     class FakeClient:
         threads = FakeThreads()
 
-    monkeypatch.setattr(thread_api, "langgraph_client", lambda: FakeClient())
+    patch_thread_module(monkeypatch, "langgraph_client", lambda: FakeClient())
 
     with pytest.raises(HTTPException) as exc_info:
-        await thread_api.proxy_dashboard_thread_run_cancel("tid", "run-1", "teammate")
+        await thread_proxy.proxy_dashboard_thread_run_cancel("tid", "run-1", "teammate")
     assert exc_info.value.status_code == 404
 
 
@@ -1361,14 +1369,14 @@ async def test_read_endpoints_accessible_by_non_owner(monkeypatch) -> None:
     class FakeClient:
         threads = FakeThreads()
 
-    monkeypatch.setattr(thread_api, "langgraph_client", lambda: FakeClient())
+    patch_thread_module(monkeypatch, "langgraph_client", lambda: FakeClient())
 
     # Read endpoints succeed for non-owners (org members).
     state = await thread_api.get_dashboard_thread_state("tid", "teammate")
     assert "values" in state
 
     # stream/events preflight should not raise.
-    await thread_api.proxy_dashboard_thread_stream_events(
+    await thread_proxy.proxy_dashboard_thread_stream_events(
         "tid", "teammate", b"{}", content_type="application/json"
     )
 
@@ -1395,17 +1403,17 @@ async def test_read_endpoints_accessible_by_non_owner(monkeypatch) -> None:
             return FakeResponse()
 
     posted: list[dict[str, object]] = []
-    monkeypatch.setattr(thread_api.httpx2, "AsyncClient", FakeAsyncClient)
-    await thread_api.proxy_dashboard_thread_history("tid", "teammate", b'{"limit": 20}')
-    await thread_api.proxy_dashboard_thread_history(
+    monkeypatch.setattr(thread_proxy.httpx2, "AsyncClient", FakeAsyncClient)
+    await thread_proxy.proxy_dashboard_thread_history("tid", "teammate", b'{"limit": 20}')
+    await thread_proxy.proxy_dashboard_thread_history(
         "tid", "teammate", b'{"limit": 20, "metadata": {"run_id": "run-1"}}'
     )
     assert posted == [
-        {"limit": thread_api._DISCOVERY_HISTORY_LIMIT},
+        {"limit": thread_proxy._DISCOVERY_HISTORY_LIMIT},
         {"limit": 20, "metadata": {"run_id": "run-1"}},
     ]
     with pytest.raises(HTTPException) as exc_info:
-        await thread_api.proxy_dashboard_thread_history("tid", "teammate", b"\xff")
+        await thread_proxy.proxy_dashboard_thread_history("tid", "teammate", b"\xff")
     assert exc_info.value.status_code == 400
 
 
@@ -1436,7 +1444,7 @@ async def test_thread_state_uses_current_run_status_when_checkpoint_is_stale(mon
         threads = FakeThreads()
         runs = FakeRuns()
 
-    monkeypatch.setattr(thread_api, "langgraph_client", lambda: FakeClient())
+    patch_thread_module(monkeypatch, "langgraph_client", lambda: FakeClient())
 
     state = await thread_api.get_dashboard_thread_state("tid", "owner")
 
@@ -1459,7 +1467,7 @@ async def test_read_endpoints_reject_non_surfaced_source(monkeypatch) -> None:
     class FakeClient:
         threads = FakeThreads()
 
-    monkeypatch.setattr(thread_api, "langgraph_client", lambda: FakeClient())
+    patch_thread_module(monkeypatch, "langgraph_client", lambda: FakeClient())
 
     with pytest.raises(HTTPException) as exc_info:
         await thread_api.get_dashboard_thread_state("tid", "owner")
@@ -1482,14 +1490,14 @@ async def test_send_dashboard_message_returns_502_when_activity_unknown(monkeypa
         assert thread_id == "tid"
         return None
 
-    monkeypatch.setattr(thread_api, "langgraph_client", lambda: FakeClient())
-    monkeypatch.setattr(thread_api, "get_thread_active_status", unknown_activity)
+    patch_thread_module(monkeypatch, "langgraph_client", lambda: FakeClient())
+    patch_thread_module(monkeypatch, "get_thread_active_status", unknown_activity)
 
     with pytest.raises(HTTPException) as exc_info:
         await thread_api.send_dashboard_message(
             "tid",
             "octocat",
-            thread_api.ThreadMessageBody(content="hello"),
+            thread_runs.ThreadMessageBody(content="hello"),
         )
 
     assert exc_info.value.status_code == 502
@@ -1514,13 +1522,13 @@ async def test_send_dashboard_message_rejects_non_admin_on_admin_thread(monkeypa
         threads = AdminThreads()
 
     monkeypatch.setenv("CONFIGURED_ADMINS", "workspace-admin")
-    monkeypatch.setattr(thread_api, "langgraph_client", lambda: AdminClient())
+    patch_thread_module(monkeypatch, "langgraph_client", lambda: AdminClient())
 
     with pytest.raises(HTTPException) as exc_info:
         await thread_api.send_dashboard_message(
             "tid",
             "teammate",
-            thread_api.ThreadMessageBody(content="ship it"),
+            thread_runs.ThreadMessageBody(content="ship it"),
         )
 
     assert exc_info.value.status_code == 403
@@ -1530,7 +1538,7 @@ async def test_send_dashboard_message_rejects_non_admin_on_admin_thread(monkeypa
 def test_assert_thread_postable_allows_configured_admin(monkeypatch) -> None:
     monkeypatch.setenv("CONFIGURED_ADMINS", "workspace-admin")
 
-    thread_api._assert_thread_postable(
+    thread_summary._assert_thread_postable(
         {"source": "dashboard", "admin_thread": True},
         "workspace-admin",
     )
@@ -1540,7 +1548,7 @@ def test_assert_thread_postable_restricts_workspace_automation(monkeypatch) -> N
     monkeypatch.setenv("CONFIGURED_ADMINS", "workspace-admin")
 
     with pytest.raises(HTTPException) as exc_info:
-        thread_api._assert_thread_postable(
+        thread_summary._assert_thread_postable(
             {"source": "schedule", "thread_category": "automation"},
             "teammate",
         )
@@ -1572,14 +1580,14 @@ async def test_send_dashboard_message_attributes_non_owner(monkeypatch) -> None:
         captured["payload"] = payload
         return True
 
-    monkeypatch.setattr(thread_api, "langgraph_client", lambda: FakeClient())
-    monkeypatch.setattr(thread_api, "get_thread_active_status", active)
-    monkeypatch.setattr(thread_api, "queue_message_for_thread", fake_queue)
+    patch_thread_module(monkeypatch, "langgraph_client", lambda: FakeClient())
+    patch_thread_module(monkeypatch, "get_thread_active_status", active)
+    patch_thread_module(monkeypatch, "queue_message_for_thread", fake_queue)
 
     await thread_api.send_dashboard_message(
         "tid",
         "teammate",
-        thread_api.ThreadMessageBody(content="ship it"),
+        thread_runs.ThreadMessageBody(content="ship it"),
     )
 
     payload = cast(dict[str, object], captured["payload"])
@@ -1610,14 +1618,14 @@ async def test_send_dashboard_message_does_not_attribute_owner(monkeypatch) -> N
         captured["payload"] = payload
         return True
 
-    monkeypatch.setattr(thread_api, "langgraph_client", lambda: FakeClient())
-    monkeypatch.setattr(thread_api, "get_thread_active_status", active)
-    monkeypatch.setattr(thread_api, "queue_message_for_thread", fake_queue)
+    patch_thread_module(monkeypatch, "langgraph_client", lambda: FakeClient())
+    patch_thread_module(monkeypatch, "get_thread_active_status", active)
+    patch_thread_module(monkeypatch, "queue_message_for_thread", fake_queue)
 
     await thread_api.send_dashboard_message(
         "tid",
         "owner",
-        thread_api.ThreadMessageBody(content="ship it"),
+        thread_runs.ThreadMessageBody(content="ship it"),
     )
 
     payload = cast(dict[str, object], captured["payload"])
@@ -1625,7 +1633,7 @@ async def test_send_dashboard_message_does_not_attribute_owner(monkeypatch) -> N
 
 
 async def test_thread_summary_exposes_resolved_state() -> None:
-    summary = await thread_api._thread_summary(
+    summary = await thread_summary._thread_summary(
         {
             "thread_id": "tid",
             "metadata": {
@@ -1642,7 +1650,7 @@ async def test_thread_summary_exposes_resolved_state() -> None:
 
 
 async def test_thread_summary_defaults_to_not_resolved() -> None:
-    summary = await thread_api._thread_summary(
+    summary = await thread_summary._thread_summary(
         {"thread_id": "tid", "metadata": {"source": "dashboard", "github_login": "octocat"}}
     )
 
@@ -1670,8 +1678,8 @@ async def test_resolve_dashboard_thread_marks_resolved(monkeypatch) -> None:
     class FakeClient:
         threads = FakeThreads()
 
-    monkeypatch.setattr(thread_api, "langgraph_client", lambda: FakeClient())
-    monkeypatch.setattr(thread_api, "agent_thread_pr_state_lock", _unlocked)
+    patch_thread_module(monkeypatch, "langgraph_client", lambda: FakeClient())
+    patch_thread_module(monkeypatch, "agent_thread_pr_state_lock", _unlocked)
 
     summary = await thread_api.resolve_dashboard_thread("tid", "octocat", resolved=True)
 
@@ -1705,8 +1713,8 @@ async def test_resolve_dashboard_thread_clears_resolved(monkeypatch) -> None:
     class FakeClient:
         threads = FakeThreads()
 
-    monkeypatch.setattr(thread_api, "langgraph_client", lambda: FakeClient())
-    monkeypatch.setattr(thread_api, "agent_thread_pr_state_lock", _unlocked)
+    patch_thread_module(monkeypatch, "langgraph_client", lambda: FakeClient())
+    patch_thread_module(monkeypatch, "agent_thread_pr_state_lock", _unlocked)
 
     summary = await thread_api.resolve_dashboard_thread("tid", "octocat", resolved=False)
 
@@ -1724,8 +1732,8 @@ async def test_resolve_dashboard_thread_rejects_unsurfaced_thread(monkeypatch) -
     class FakeClient:
         threads = FakeThreads()
 
-    monkeypatch.setattr(thread_api, "langgraph_client", lambda: FakeClient())
-    monkeypatch.setattr(thread_api, "agent_thread_pr_state_lock", _unlocked)
+    patch_thread_module(monkeypatch, "langgraph_client", lambda: FakeClient())
+    patch_thread_module(monkeypatch, "agent_thread_pr_state_lock", _unlocked)
 
     with pytest.raises(HTTPException) as exc_info:
         await thread_api.resolve_dashboard_thread("tid", "teammate", resolved=True)
@@ -1756,13 +1764,13 @@ async def test_enrich_run_start_command_unresolves_thread(monkeypatch) -> None:
         threads = FakeThreads()
 
     _patch_new_thread_deps(monkeypatch, profile={})
-    monkeypatch.setattr(thread_api, "langgraph_client", lambda: FakeClient())
-    monkeypatch.setattr(thread_api, "agent_thread_pr_state_lock", _unlocked)
+    patch_thread_module(monkeypatch, "langgraph_client", lambda: FakeClient())
+    patch_thread_module(monkeypatch, "agent_thread_pr_state_lock", _unlocked)
 
     async def fake_build(thread_id, login, metadata, *, overrides):
         return {"github_login": login, "source": "dashboard"}
 
-    monkeypatch.setattr(thread_api, "_build_dashboard_configurable", fake_build)
+    patch_thread_module(monkeypatch, "_build_dashboard_configurable", fake_build)
 
     command = {
         "method": "run.start",
@@ -1772,7 +1780,7 @@ async def test_enrich_run_start_command_unresolves_thread(monkeypatch) -> None:
         },
     }
 
-    await thread_api._enrich_run_start_command(
+    await thread_runs._enrich_run_start_command(
         "tid",
         "octocat",
         command,
@@ -1804,19 +1812,19 @@ def test_summary_matches_filters() -> None:
         "pr": {"number": 123, "url": "https://github.com/langchain-ai/open-swe/pull/123"},
     }
 
-    assert thread_api._summary_matches_filters(
+    assert thread_listing._summary_matches_filters(
         summary, resolved=True, viewed=None, source=None, status=None, query=None
     )
-    assert not thread_api._summary_matches_filters(
+    assert not thread_listing._summary_matches_filters(
         summary, resolved=False, viewed=None, source=None, status=None, query=None
     )
-    assert thread_api._summary_matches_filters(
+    assert thread_listing._summary_matches_filters(
         summary, resolved=None, viewed=None, source="github", status=None, query="flaky"
     )
-    assert not thread_api._summary_matches_filters(
+    assert not thread_listing._summary_matches_filters(
         summary, resolved=None, viewed=None, source=None, status=None, query="missing"
     )
-    assert thread_api._summary_matches_filters(
+    assert thread_listing._summary_matches_filters(
         summary, resolved=None, viewed=None, source=None, status=None, query="pull/123"
     )
 
@@ -1843,27 +1851,29 @@ def test_metadata_matches_filters() -> None:
         "title": "Scheduled cleanup",
     }
 
-    assert thread_api._metadata_matches_filters(metadata, resolved=True, source=None, query=None)
-    assert not thread_api._metadata_matches_filters(
+    assert thread_listing._metadata_matches_filters(
+        metadata, resolved=True, source=None, query=None
+    )
+    assert not thread_listing._metadata_matches_filters(
         metadata, resolved=False, source=None, query=None
     )
-    assert thread_api._metadata_matches_filters(
+    assert thread_listing._metadata_matches_filters(
         metadata, resolved=None, source="dashboard", query="login"
     )
-    assert not thread_api._metadata_matches_filters(
+    assert not thread_listing._metadata_matches_filters(
         metadata, resolved=None, source="github", query=None
     )
     for query in ("langchain-ai/open-swe", "fix-login", "pull/123"):
-        assert thread_api._metadata_matches_filters(
+        assert thread_listing._metadata_matches_filters(
             metadata, resolved=None, source=None, query=query
         )
-    assert thread_api._metadata_matches_filters(
+    assert thread_listing._metadata_matches_filters(
         metadata, resolved=None, source=None, query=None, scope="interactive"
     )
-    assert not thread_api._metadata_matches_filters(
+    assert not thread_listing._metadata_matches_filters(
         automation, resolved=None, source=None, query=None, scope="interactive"
     )
-    assert thread_api._metadata_matches_filters(
+    assert thread_listing._metadata_matches_filters(
         automation,
         resolved=None,
         source=None,
@@ -1871,7 +1881,7 @@ def test_metadata_matches_filters() -> None:
         scope="automation",
         automation_id="schedule-1",
     )
-    assert not thread_api._metadata_matches_filters(
+    assert not thread_listing._metadata_matches_filters(
         automation,
         resolved=None,
         source=None,
@@ -1900,7 +1910,7 @@ def _make_threads(count: int, *, resolved_before: int) -> list[dict[str, object]
 
 
 async def test_list_dashboard_threads_page_pages_beyond_first_search_batch(monkeypatch) -> None:
-    page_size = thread_api._THREADS_SEARCH_PAGE
+    page_size = thread_listing._THREADS_SEARCH_PAGE
     threads = _make_threads(page_size + 50, resolved_before=page_size)
     for thread in threads:
         cast(dict[str, object], thread["metadata"])["latest_run_status"] = "success"
@@ -1910,7 +1920,7 @@ async def test_list_dashboard_threads_page_pages_beyond_first_search_batch(monke
     class FakeThreads:
         async def search(self, *, metadata, limit, offset, sort_by, sort_order, select):
             offsets.append(offset)
-            assert select == thread_api._THREAD_LIST_SELECT
+            assert select == thread_listing._THREAD_LIST_SELECT
             return threads[offset : offset + limit]
 
         async def update(self, *, thread_id, metadata):
@@ -1926,9 +1936,9 @@ async def test_list_dashboard_threads_page_pages_beyond_first_search_batch(monke
         threads = FakeThreads()
         runs = FakeRuns()
 
-    monkeypatch.setattr(thread_api, "langgraph_client", lambda: FakeClient())
+    patch_thread_module(monkeypatch, "langgraph_client", lambda: FakeClient())
 
-    result = await thread_api.list_dashboard_threads_page(
+    result = await thread_listing.list_dashboard_threads_page(
         "octocat", email=None, limit=25, offset=0, resolved=False
     )
 
@@ -1974,12 +1984,12 @@ async def test_list_dashboard_threads_page_scopes_automation_runs(monkeypatch) -
         threads = FakeThreads()
         runs = FakeRuns()
 
-    monkeypatch.setattr(thread_api, "langgraph_client", lambda: FakeClient())
+    patch_thread_module(monkeypatch, "langgraph_client", lambda: FakeClient())
 
-    interactive = await thread_api.list_dashboard_threads_page(
+    interactive = await thread_listing.list_dashboard_threads_page(
         "octocat", email=None, scope="interactive"
     )
-    automation = await thread_api.list_dashboard_threads_page(
+    automation = await thread_listing.list_dashboard_threads_page(
         "octocat", email=None, scope="automation", automation_id="schedule-1"
     )
 
@@ -2029,9 +2039,9 @@ async def test_list_dashboard_threads_page_scopes_search_to_requested_participan
         threads = FakeThreads()
         runs = FakeRuns()
 
-    monkeypatch.setattr(thread_api, "langgraph_client", lambda: FakeClient())
+    patch_thread_module(monkeypatch, "langgraph_client", lambda: FakeClient())
 
-    result = await thread_api.list_dashboard_threads_page(
+    result = await thread_listing.list_dashboard_threads_page(
         "admin-user",
         email="admin@example.com",
         filter_participant_login="other-user",
@@ -2066,13 +2076,13 @@ async def test_list_dashboard_threads_page_filters_ownerless_threads(monkeypatch
         async def list(self, thread_id, limit=1):
             return []
 
-    monkeypatch.setattr(
-        thread_api,
+    patch_thread_module(
+        monkeypatch,
         "langgraph_client",
         lambda: SimpleNamespace(threads=FakeThreads(), runs=FakeRuns()),
     )
 
-    result = await thread_api.list_dashboard_threads_page("octocat", email=None, ownerless=True)
+    result = await thread_listing.list_dashboard_threads_page("octocat", email=None, ownerless=True)
 
     assert [item["id"] for item in result["items"]] == ["t2"]
 
@@ -2112,13 +2122,13 @@ async def test_list_dashboard_threads_page_filters_flat_and_legacy_repo_metadata
         async def list(self, thread_id, limit=1):
             return []
 
-    monkeypatch.setattr(
-        thread_api,
+    patch_thread_module(
+        monkeypatch,
         "langgraph_client",
         lambda: SimpleNamespace(threads=FakeThreads(), runs=FakeRuns()),
     )
 
-    result = await thread_api.list_dashboard_threads_page(
+    result = await thread_listing.list_dashboard_threads_page(
         "octocat", email=None, repo="langchain-ai/open-swe"
     )
 
@@ -2156,13 +2166,13 @@ async def test_list_dashboard_thread_projects_discovers_metadata_without_summari
         async def search(self, *, metadata, limit, offset, sort_by, sort_order, select):
             return threads[offset : offset + limit]
 
-    monkeypatch.setattr(
-        thread_api,
+    patch_thread_module(
+        monkeypatch,
         "langgraph_client",
         lambda: SimpleNamespace(threads=FakeThreads()),
     )
 
-    result = await thread_api.list_dashboard_thread_projects("octocat")
+    result = await thread_listing.list_dashboard_thread_projects("octocat")
 
     assert result == [
         {
@@ -2188,8 +2198,8 @@ async def test_pin_dashboard_thread_allows_readable_non_owner(monkeypatch) -> No
                 "metadata": {"source": "slack", "github_login": "other"},
             }
 
-    monkeypatch.setattr(
-        thread_api,
+    patch_thread_module(
+        monkeypatch,
         "langgraph_client",
         lambda: SimpleNamespace(threads=FakeThreads()),
     )
@@ -2197,9 +2207,9 @@ async def test_pin_dashboard_thread_allows_readable_non_owner(monkeypatch) -> No
     async def fake_pin(login: str, thread_id: str) -> None:
         pinned.append((login, thread_id))
 
-    monkeypatch.setattr(thread_api, "pin_thread", fake_pin)
+    patch_thread_module(monkeypatch, "pin_thread", fake_pin)
 
-    await thread_api.pin_dashboard_thread("shared-thread", "octocat")
+    await thread_listing.pin_dashboard_thread("shared-thread", "octocat")
 
     assert pinned == [("octocat", "shared-thread")]
 
@@ -2209,14 +2219,14 @@ async def test_pin_dashboard_thread_returns_not_found_for_lookup_failure(monkeyp
         async def get(self, thread_id):
             raise RuntimeError("thread missing")
 
-    monkeypatch.setattr(
-        thread_api,
+    patch_thread_module(
+        monkeypatch,
         "langgraph_client",
         lambda: SimpleNamespace(threads=FakeThreads()),
     )
 
     with pytest.raises(HTTPException) as exc_info:
-        await thread_api.pin_dashboard_thread("missing-thread", "octocat")
+        await thread_listing.pin_dashboard_thread("missing-thread", "octocat")
 
     assert exc_info.value.status_code == 404
     assert exc_info.value.detail == "thread not found"
@@ -2227,14 +2237,14 @@ async def test_pin_dashboard_thread_rejects_unreadable_thread(monkeypatch) -> No
         async def get(self, thread_id):
             return {"thread_id": thread_id, "metadata": {"source": "internal"}}
 
-    monkeypatch.setattr(
-        thread_api,
+    patch_thread_module(
+        monkeypatch,
         "langgraph_client",
         lambda: SimpleNamespace(threads=FakeThreads()),
     )
 
     with pytest.raises(HTTPException) as exc_info:
-        await thread_api.pin_dashboard_thread("private-thread", "octocat")
+        await thread_listing.pin_dashboard_thread("private-thread", "octocat")
 
     assert exc_info.value.status_code == 404
 
@@ -2269,8 +2279,8 @@ async def test_list_dashboard_pinned_threads_returns_only_readable_threads(
         async def list(self, thread_id, limit=1):
             return []
 
-    monkeypatch.setattr(
-        thread_api,
+    patch_thread_module(
+        monkeypatch,
         "langgraph_client",
         lambda: SimpleNamespace(threads=FakeThreads(), runs=FakeRuns()),
     )
@@ -2279,9 +2289,9 @@ async def test_list_dashboard_pinned_threads_returns_only_readable_threads(
         assert login == "octocat"
         return ["shared-thread", "private-thread", "missing-thread"]
 
-    monkeypatch.setattr(thread_api, "list_thread_pin_ids", fake_pin_ids)
+    patch_thread_module(monkeypatch, "list_thread_pin_ids", fake_pin_ids)
 
-    result = await thread_api.list_dashboard_pinned_threads("octocat")
+    result = await thread_listing.list_dashboard_pinned_threads("octocat")
 
     assert [item["id"] for item in result] == ["shared-thread"]
 
@@ -2316,9 +2326,9 @@ async def test_list_dashboard_threads_page_can_sort_by_creation_time(monkeypatch
         threads = FakeThreads()
         runs = FakeRuns()
 
-    monkeypatch.setattr(thread_api, "langgraph_client", lambda: FakeClient())
+    patch_thread_module(monkeypatch, "langgraph_client", lambda: FakeClient())
 
-    result = await thread_api.list_dashboard_threads_page(
+    result = await thread_listing.list_dashboard_threads_page(
         "octocat", email=None, limit=2, offset=0, sort_by="created_at"
     )
 
@@ -2350,9 +2360,11 @@ async def test_list_dashboard_threads_page_refreshes_only_unsettled_threads(monk
         threads = FakeThreads()
         runs = FakeRuns()
 
-    monkeypatch.setattr(thread_api, "langgraph_client", lambda: FakeClient())
+    patch_thread_module(monkeypatch, "langgraph_client", lambda: FakeClient())
 
-    result = await thread_api.list_dashboard_threads_page("octocat", email=None, limit=3, offset=0)
+    result = await thread_listing.list_dashboard_threads_page(
+        "octocat", email=None, limit=3, offset=0
+    )
 
     assert run_list_thread_ids == ["t1"]
     assert updates == [
@@ -2387,9 +2399,9 @@ async def test_status_filter_refreshes_threads_missing_run_status(monkeypatch) -
         threads = FakeThreads()
         runs = FakeRuns()
 
-    monkeypatch.setattr(thread_api, "langgraph_client", lambda: FakeClient())
+    patch_thread_module(monkeypatch, "langgraph_client", lambda: FakeClient())
 
-    result = await thread_api.list_dashboard_threads_page(
+    result = await thread_listing.list_dashboard_threads_page(
         "octocat", email=None, limit=25, offset=0, status="finished"
     )
 
@@ -2506,9 +2518,9 @@ async def test_working_tree_diff_reads_live_sandbox_against_head(monkeypatch) ->
         "truncated": False,
         "summary": {"files": 1, "additions": 1, "deletions": 0},
     }
-    monkeypatch.setattr(thread_api, "_readable_thread_metadata", AsyncMock(return_value=metadata))
+    patch_thread_module(monkeypatch, "_readable_thread_metadata", AsyncMock(return_value=metadata))
     sandbox = object()
-    monkeypatch.setattr(thread_api, "create_sandbox", AsyncMock(return_value=sandbox))
+    patch_thread_module(monkeypatch, "create_sandbox", AsyncMock(return_value=sandbox))
     monkeypatch.setattr(
         "agent.sandboxes.paths.resolve_sandbox_work_dir",
         AsyncMock(return_value="/work"),
@@ -2516,7 +2528,7 @@ async def test_working_tree_diff_reads_live_sandbox_against_head(monkeypatch) ->
     read_diff = AsyncMock(return_value=live)
     monkeypatch.setattr("agent.utils.turn_checkpoint.read_turn_diff", read_diff)
 
-    result = await thread_api.get_dashboard_thread_working_tree_diff("thread-1", "owner")
+    result = await thread_diffs.get_dashboard_thread_working_tree_diff("thread-1", "owner")
 
     assert result == live
     read_diff.assert_awaited_once_with(sandbox, "/work", "HEAD", None, repo_path="/work/repo")
@@ -2524,11 +2536,11 @@ async def test_working_tree_diff_reads_live_sandbox_against_head(monkeypatch) ->
 
 async def test_working_tree_diff_raises_when_the_sandbox_is_unreachable(monkeypatch) -> None:
     metadata = {"sandbox_id": "sandbox-1"}
-    monkeypatch.setattr(thread_api, "_readable_thread_metadata", AsyncMock(return_value=metadata))
-    monkeypatch.setattr(thread_api, "create_sandbox", AsyncMock(side_effect=RuntimeError))
+    patch_thread_module(monkeypatch, "_readable_thread_metadata", AsyncMock(return_value=metadata))
+    patch_thread_module(monkeypatch, "create_sandbox", AsyncMock(side_effect=RuntimeError))
 
     with pytest.raises(HTTPException) as exc_info:
-        await thread_api.get_dashboard_thread_working_tree_diff("thread-1", "owner")
+        await thread_diffs.get_dashboard_thread_working_tree_diff("thread-1", "owner")
 
     assert exc_info.value.status_code == 503
     assert exc_info.value.detail == "Could not connect to the workspace."
@@ -2541,14 +2553,14 @@ async def test_branch_diff_uses_repository_from_pr_url(monkeypatch) -> None:
         "pr_number": 1925,
         "pr_url": "https://github.com/langchain-ai/open-swe/pull/1925",
     }
-    monkeypatch.setattr(thread_api, "_readable_thread_metadata", AsyncMock(return_value=metadata))
-    monkeypatch.setattr(thread_api, "_github_token_for_login", AsyncMock(return_value="token"))
+    patch_thread_module(monkeypatch, "_readable_thread_metadata", AsyncMock(return_value=metadata))
+    patch_thread_module(monkeypatch, "_github_token_for_login", AsyncMock(return_value="token"))
     build_diff = AsyncMock(
         return_value={"base_sha": "base", "head_sha": "head", "truncated": False, "files": []}
     )
-    monkeypatch.setattr(thread_api, "build_pr_diff_files", build_diff)
+    patch_thread_module(monkeypatch, "build_pr_diff_files", build_diff)
 
-    await thread_api.get_dashboard_thread_branch_diff("thread-1", "owner")
+    await thread_diffs.get_dashboard_thread_branch_diff("thread-1", "owner")
 
     assert build_diff.await_args is not None
     assert build_diff.await_args.args[1:] == ("langchain-ai/open-swe", 1925)
@@ -2561,14 +2573,14 @@ async def test_branch_diff_without_a_pull_request_compares_against_the_base(monk
         "base_branch": "main",
         "branch_name": "open-swe/feature",
     }
-    monkeypatch.setattr(thread_api, "_readable_thread_metadata", AsyncMock(return_value=metadata))
-    monkeypatch.setattr(thread_api, "_github_token_for_login", AsyncMock(return_value="token"))
+    patch_thread_module(monkeypatch, "_readable_thread_metadata", AsyncMock(return_value=metadata))
+    patch_thread_module(monkeypatch, "_github_token_for_login", AsyncMock(return_value="token"))
     build_compare = AsyncMock(
         return_value={"base_sha": "merge-base", "head_sha": "head", "truncated": False, "files": []}
     )
-    monkeypatch.setattr(thread_api, "build_compare_diff_files", build_compare)
+    patch_thread_module(monkeypatch, "build_compare_diff_files", build_compare)
 
-    result = await thread_api.get_dashboard_thread_branch_diff("thread-1", "owner")
+    result = await thread_diffs.get_dashboard_thread_branch_diff("thread-1", "owner")
 
     assert build_compare.await_args is not None
     assert build_compare.await_args.args[1:] == (
@@ -2587,13 +2599,13 @@ async def test_branch_diff_rejects_an_unsafe_branch_name(monkeypatch) -> None:
         "base_branch": "main",
         "branch_name": "../../etc/passwd",
     }
-    monkeypatch.setattr(thread_api, "_readable_thread_metadata", AsyncMock(return_value=metadata))
-    monkeypatch.setattr(thread_api, "_github_token_for_login", AsyncMock(return_value="token"))
+    patch_thread_module(monkeypatch, "_readable_thread_metadata", AsyncMock(return_value=metadata))
+    patch_thread_module(monkeypatch, "_github_token_for_login", AsyncMock(return_value="token"))
     build_compare = AsyncMock()
-    monkeypatch.setattr(thread_api, "build_compare_diff_files", build_compare)
+    patch_thread_module(monkeypatch, "build_compare_diff_files", build_compare)
 
     with pytest.raises(HTTPException) as excinfo:
-        await thread_api.get_dashboard_thread_branch_diff("thread-1", "owner")
+        await thread_diffs.get_dashboard_thread_branch_diff("thread-1", "owner")
 
     assert excinfo.value.status_code == 404
     build_compare.assert_not_awaited()
@@ -2640,7 +2652,7 @@ async def test_cancel_dashboard_thread_interrupts_runs_it_did_not_start(monkeypa
         runs = FakeRuns()
         store = FakeStore()
 
-    monkeypatch.setattr(thread_api, "langgraph_client", lambda: FakeClient())
+    patch_thread_module(monkeypatch, "langgraph_client", lambda: FakeClient())
 
     result = await thread_api.cancel_dashboard_thread("thread-1", "owner")
 
@@ -2679,7 +2691,7 @@ async def test_cancel_dashboard_thread_rejects_non_owner(monkeypatch) -> None:
         threads = FakeThreads()
         runs = FakeRuns()
 
-    monkeypatch.setattr(thread_api, "langgraph_client", lambda: FakeClient())
+    patch_thread_module(monkeypatch, "langgraph_client", lambda: FakeClient())
 
     with pytest.raises(HTTPException):
         await thread_api.cancel_dashboard_thread("thread-1", "someone-else")
@@ -2722,7 +2734,7 @@ async def test_admin_cancel_dashboard_thread_interrupts_all_active_runs(monkeypa
         threads = FakeThreads()
         runs = FakeRuns()
 
-    monkeypatch.setattr(thread_api, "langgraph_client", lambda: FakeClient())
+    patch_thread_module(monkeypatch, "langgraph_client", lambda: FakeClient())
 
     result = await thread_api.admin_cancel_dashboard_thread("thread-1")
 
@@ -2761,7 +2773,7 @@ async def test_admin_cancel_dashboard_thread_does_not_update_on_cancel_failure(m
         threads = FakeThreads()
         runs = FakeRuns()
 
-    monkeypatch.setattr(thread_api, "langgraph_client", lambda: FakeClient())
+    patch_thread_module(monkeypatch, "langgraph_client", lambda: FakeClient())
 
     with pytest.raises(HTTPException) as exc_info:
         await thread_api.admin_cancel_dashboard_thread("thread-1")
