+++ tests/e2e/patches.py
@@ -75,14 +75,15 @@ def apply() -> None:
     # OAuth-token store is an external credential boundary. Stub it so a web
     # follow-up (dashboard run.start) and PR-as-user resolution have a token;
     # the real ownership/authorization checks still run.
-    from agent.dashboard import profiles, thread_api
+    from agent.dashboard import profiles
+    from agent.dashboard.threads import access as thread_access
     from agent.github import pull_request_context, pull_request_status
 
     async def _dummy_user_token(login: str, **_kwargs: object) -> str:  # noqa: ARG001
         return "dummy-user-oauth-token"
 
     profiles.get_valid_access_token = _dummy_user_token
-    thread_api.get_valid_access_token = _dummy_user_token
+    thread_access.get_valid_access_token = _dummy_user_token
     pull_request_status.GITHUB_API_BASE = FAKE_GITHUB_API
     pull_request_status.GITHUB_GRAPHQL = f"{FAKE_GITHUB_API}/graphql"
     pull_request_context.GITHUB_GRAPHQL = f"{FAKE_GITHUB_API}/graphql"
